Fix playwright hooks
